pyproject.toml
[tool.poetry]
name = "hmip2mqtt"
version = "0.1.0"
description = ""
authors = ["Markus Mohnen <markus.mohnen@gmail.com>"]

[tool.poetry.scripts]
hmip2mqtt = "hmip2mqtt.cli:main"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.3.2"
appdirs = "^1.4.4"
homematicip = "^0.13.1"
paho-mqtt = "^1.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
